+++ cads_adaptors/__init__.py
@@ -23,7 +23,7 @@ except ImportError:  # pragma: no cover
     __version__ = "999"
 
 from cads_adaptors.adaptors import AbstractAdaptor, Context, DummyAdaptor
-from cads_adaptors.adaptors.cadsobs import ObservationsAdaptor
+from cads_adaptors.adaptors.cadsobs.adaptor import ObservationsAdaptor
 from cads_adaptors.adaptors.cds import AbstractCdsAdaptor, DummyCdsAdaptor
 from cads_adaptors.adaptors.insitu import (
     InsituDatabaseCdsAdaptor,
@@ -53,6 +53,6 @@ __all__ = [
     "UrlCdsAdaptor",
     "MultiAdaptor",
     "MultiMarsCdsAdaptor",
-    "ObservationsAdaptor",
     "RoocsCdsAdaptor",
+    "ObservationsAdaptor",
 ]
